Projeto finalizado

+++ src/app/componentes/listavisitas/listavisitas.component.ts
@@ -4,7 +4,7 @@ import {ListavisitaService} from '../../services/listavisita.service';
 import {Listas} from './Listas'
 import {Pontos} from '../pontos-turisticos/Pontos'
 import { NgForm } from '@angular/forms';
-
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-listavisitas',
@@ -19,19 +19,22 @@ lista: Listas [];
 pontos = {} as Pontos;
 ponto: Pontos[];
 
-mostrar: boolean = false;
-
+formVisibility= false;
 
 
   constructor(private route: ActivatedRoute, private listaServ: ListavisitaService) { 
-    this.mostrar = false;
+    this.formVisibility= false;
   }
 
   ngOnInit(): void {
     this.getListaVisita();
 
   }
-
+  
+  ShowForm(){
+    this.formVisibility=true;
+    console.log(this.formVisibility);
+  }
   getPontosTuristicos(){
     this.listaServ.getTodosPontosTuristicos()
     .subscribe(dados => {
@@ -57,7 +60,10 @@ mostrar: boolean = false;
     if (this.listas.id !== undefined) {
       this.listaServ.updateListaVisita(this.listas).subscribe(() => {
         this.cleanForm(form);
+        Swal.fire('Obrigada','Dados atualizados corretamente','success' )
+        this.formVisibility=false;
         this.getListaVisita();
+
       });
     } else {
       this.listaServ.createListaVisita(this.listas).subscribe(() => {
@@ -73,20 +79,51 @@ mostrar: boolean = false;
     form.resetForm();
   }
 
-  Mostrar(){
-    this.mostrar = true;
-  }
+  
 
   editLista(Lista: Listas) {
-    
-    this.listas = { ...Lista };
+    Swal.fire({
+      title: 'Deseja editar seu registro?',
+      text: '',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Sim, desejo editar',
+      cancelButtonText: 'Não, não quero editar',
+    }).then((result) => {
+      if (result.isConfirmed){
+        this.formVisibility=true;
+        console.log(this.formVisibility)
+        this.listas = { ...Lista };
+      }
+      else {
+        Swal.fire('Cancelado', 'Operação Cancelada','error')
+      }
+
+    })
     
   }
 
   deleteLista(Lista: Listas) {
-    this.listaServ.deleteListaVisita(Lista).subscribe(() => {
-      this.getListaVisita();
-    });
+    Swal.fire({
+      title: 'Deseja deletar seu registro?',
+      text: '',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Sim, pode deletar',
+      cancelButtonText: 'Não, não quero deletar',
+    }).then((result) => {
+      if (result.isConfirmed){
+        this.listaServ.deleteListaVisita(Lista).subscribe(() => {
+          this.getListaVisita();
+          Swal.fire('Obrigada','Dados deletados corretamente','success' )
+        });
+      }
+      else {
+        Swal.fire('Cancelado', 'Operação Cancelada','error')
+      }
+
+    })
+    
   }
 
 }
